Git Issue #10 Complete
The main points of git issue #10 are completed.
Namely, compiling Fluigi will generate Job objects, and the Job object
id's will be collected under the user object. The next commit will
adress:
-> Knockout UI for the "jobs" or "solutions"
-> A crash issue that occures on the first run of Fluigi

Otherwise, moving on to the full-clickthrough, SVG previews, downloads
for eps and svg, and and general bugfixes and UI cleanups.