Continued work on download... still having trouble
Generally I am able to get the file content and information from the client to the server when the associated file is clicked. I am also able to create zp files, which I'll soon expand to allowing a zip download of a workspace.

This being said, the browser does not automatically prompt the download, and I have to maunally open it through the web console >.>